requirements.txt
moviepy
gTTS
python-pptx
nltk
pyttsx3  # Optional, if using offline TTS
